requirements.txt
arrow==1.3.0
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
urllib3==2.1.0
torch>=2.1.0
timm>=0.9.12
pytorch-lightning>=2.1.3
dvc==3.38.1
dvc-gdrive==3.0.1
dvc-gs==3.0.1
hydra-core==1.3.2
omegaconf==2.3.0
numpy==1.26.3
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2
matplotlib==3.8.2
wandb==0.16.2
pytest==7.4.4
pre-commit>=3.6.0
